Export prompb.CompareLower
This was duplicated in another packages that already depended
on the prompb package.
